LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/rotate-interpolation-expected.txt


FAIL CSS Transitions: property <rotate> from [100deg] to [180deg] at (-1) should be [20deg] assert_equals: expected "20deg " but got "180deg "
FAIL CSS Transitions: property <rotate> from [100deg] to [180deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "180deg "
FAIL CSS Transitions: property <rotate> from [100deg] to [180deg] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "180deg "
FAIL CSS Transitions: property <rotate> from [100deg] to [180deg] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "180deg "
PASS CSS Transitions: property <rotate> from [100deg] to [180deg] at (1) should be [180deg]
FAIL CSS Transitions: property <rotate> from [100deg] to [180deg] at (2) should be [260deg] assert_equals: expected "260deg " but got "180deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [180deg] at (-1) should be [20deg] assert_equals: expected "20deg " but got "180deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [180deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "180deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [180deg] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "180deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [180deg] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "180deg "
PASS CSS Transitions with transition: all: property <rotate> from [100deg] to [180deg] at (1) should be [180deg]
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [180deg] at (2) should be [260deg] assert_equals: expected "260deg " but got "180deg "
FAIL CSS Animations: property <rotate> from [100deg] to [180deg] at (-1) should be [20deg] assert_equals: expected "20deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [180deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [180deg] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [180deg] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [180deg] at (1) should be [180deg] assert_equals: expected "180deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [180deg] at (2) should be [260deg] assert_equals: expected "260deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [180deg] at (-1) should be [20deg] assert_equals: expected "20deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [180deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [180deg] at (0.125) should be [110deg] assert_equals: expected "110deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [180deg] at (0.875) should be [170deg] assert_equals: expected "170deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [180deg] at (1) should be [180deg] assert_equals: expected "180deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [180deg] at (2) should be [260deg] assert_equals: expected "260deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [45deg] to [-1 1 0 60deg] at (-1) should be [0.447214 -0.447214 0.774597 104.478deg] assert_equals: expected "0.45 - 0.45 0.77 104.48deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions: property <rotate> from [45deg] to [-1 1 0 60deg] at (0) should be [z 45deg] assert_equals: expected "45deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.125) should be [-0.136456 0.136456 0.981203 40.6037deg] assert_equals: expected "- 0.14 0.14 0.98 40.6deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.875) should be [-0.70246 0.70246 0.114452 53.1994deg] assert_equals: expected "- 0.7 0.7 0.11 53.2deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions: property <rotate> from [45deg] to [-1 1 0 60deg] at (1) should be [-0.71 0.71 0 60deg] assert_equals: expected "- 0.71 0.71 0 60deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions: property <rotate> from [45deg] to [-1 1 0 60deg] at (2) should be [-0.637897 0.637897 -0.431479 124.975deg] assert_equals: expected "- 0.64 0.64 - 0.43 124.97deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions with transition: all: property <rotate> from [45deg] to [-1 1 0 60deg] at (-1) should be [0.447214 -0.447214 0.774597 104.478deg] assert_equals: expected "0.45 - 0.45 0.77 104.48deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions with transition: all: property <rotate> from [45deg] to [-1 1 0 60deg] at (0) should be [z 45deg] assert_equals: expected "45deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions with transition: all: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.125) should be [-0.136456 0.136456 0.981203 40.6037deg] assert_equals: expected "- 0.14 0.14 0.98 40.6deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions with transition: all: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.875) should be [-0.70246 0.70246 0.114452 53.1994deg] assert_equals: expected "- 0.7 0.7 0.11 53.2deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions with transition: all: property <rotate> from [45deg] to [-1 1 0 60deg] at (1) should be [-0.71 0.71 0 60deg] assert_equals: expected "- 0.71 0.71 0 60deg " but got "- 1 1 0 60deg "
FAIL CSS Transitions with transition: all: property <rotate> from [45deg] to [-1 1 0 60deg] at (2) should be [-0.637897 0.637897 -0.431479 124.975deg] assert_equals: expected "- 0.64 0.64 - 0.43 124.97deg " but got "- 1 1 0 60deg "
FAIL CSS Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (-1) should be [0.447214 -0.447214 0.774597 104.478deg] assert_equals: expected "0.45 - 0.45 0.77 104.48deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (0) should be [z 45deg] assert_equals: expected "45deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.125) should be [-0.136456 0.136456 0.981203 40.6037deg] assert_equals: expected "- 0.14 0.14 0.98 40.6deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.875) should be [-0.70246 0.70246 0.114452 53.1994deg] assert_equals: expected "- 0.7 0.7 0.11 53.2deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (1) should be [-0.71 0.71 0 60deg] assert_equals: expected "- 0.71 0.71 0 60deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (2) should be [-0.637897 0.637897 -0.431479 124.975deg] assert_equals: expected "- 0.64 0.64 - 0.43 124.97deg " but got "10deg "
FAIL Web Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (-1) should be [0.447214 -0.447214 0.774597 104.478deg] assert_equals: expected "0.45 - 0.45 0.77 104.48deg " but got "10deg "
FAIL Web Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (0) should be [z 45deg] assert_equals: expected "45deg " but got "10deg "
FAIL Web Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.125) should be [-0.136456 0.136456 0.981203 40.6037deg] assert_equals: expected "- 0.14 0.14 0.98 40.6deg " but got "10deg "
FAIL Web Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (0.875) should be [-0.70246 0.70246 0.114452 53.1994deg] assert_equals: expected "- 0.7 0.7 0.11 53.2deg " but got "10deg "
FAIL Web Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (1) should be [-0.71 0.71 0 60deg] assert_equals: expected "- 0.71 0.71 0 60deg " but got "10deg "
FAIL Web Animations: property <rotate> from [45deg] to [-1 1 0 60deg] at (2) should be [-0.637897 0.637897 -0.431479 124.975deg] assert_equals: expected "- 0.64 0.64 - 0.43 124.97deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [none] to [7 -8 9 400grad] at (-1) should be [0.5 -0.57 0.65 -400grad] assert_equals: expected "0.5 - 0.57 0.65 - 360deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions: property <rotate> from [none] to [7 -8 9 400grad] at (0) should be [0.5 -0.57 0.65 0deg] assert_equals: expected "0.5 - 0.57 0.65 0deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions: property <rotate> from [none] to [7 -8 9 400grad] at (0.125) should be [0.5 -0.57 0.65 50grad] assert_equals: expected "0.5 - 0.57 0.65 45deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions: property <rotate> from [none] to [7 -8 9 400grad] at (0.875) should be [0.5 -0.57 0.65 350grad] assert_equals: expected "0.5 - 0.57 0.65 315deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions: property <rotate> from [none] to [7 -8 9 400grad] at (1) should be [0.5 -0.57 0.65 400grad] assert_equals: expected "0.5 - 0.57 0.65 360deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions: property <rotate> from [none] to [7 -8 9 400grad] at (2) should be [0.5 -0.57 0.65 800grad] assert_equals: expected "0.5 - 0.57 0.65 720deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [7 -8 9 400grad] at (-1) should be [0.5 -0.57 0.65 -400grad] assert_equals: expected "0.5 - 0.57 0.65 - 360deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [7 -8 9 400grad] at (0) should be [0.5 -0.57 0.65 0deg] assert_equals: expected "0.5 - 0.57 0.65 0deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [7 -8 9 400grad] at (0.125) should be [0.5 -0.57 0.65 50grad] assert_equals: expected "0.5 - 0.57 0.65 45deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [7 -8 9 400grad] at (0.875) should be [0.5 -0.57 0.65 350grad] assert_equals: expected "0.5 - 0.57 0.65 315deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [7 -8 9 400grad] at (1) should be [0.5 -0.57 0.65 400grad] assert_equals: expected "0.5 - 0.57 0.65 360deg " but got "7 - 8 9 360deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [7 -8 9 400grad] at (2) should be [0.5 -0.57 0.65 800grad] assert_equals: expected "0.5 - 0.57 0.65 720deg " but got "7 - 8 9 360deg "
FAIL CSS Animations: property <rotate> from [none] to [7 -8 9 400grad] at (-1) should be [0.5 -0.57 0.65 -400grad] assert_equals: expected "0.5 - 0.57 0.65 - 360deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [7 -8 9 400grad] at (0) should be [0.5 -0.57 0.65 0deg] assert_equals: expected "0.5 - 0.57 0.65 0deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [7 -8 9 400grad] at (0.125) should be [0.5 -0.57 0.65 50grad] assert_equals: expected "0.5 - 0.57 0.65 45deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [7 -8 9 400grad] at (0.875) should be [0.5 -0.57 0.65 350grad] assert_equals: expected "0.5 - 0.57 0.65 315deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [7 -8 9 400grad] at (1) should be [0.5 -0.57 0.65 400grad] assert_equals: expected "0.5 - 0.57 0.65 360deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [7 -8 9 400grad] at (2) should be [0.5 -0.57 0.65 800grad] assert_equals: expected "0.5 - 0.57 0.65 720deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [7 -8 9 400grad] at (-1) should be [0.5 -0.57 0.65 -400grad] assert_equals: expected "0.5 - 0.57 0.65 - 360deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [7 -8 9 400grad] at (0) should be [0.5 -0.57 0.65 0deg] assert_equals: expected "0.5 - 0.57 0.65 0deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [7 -8 9 400grad] at (0.125) should be [0.5 -0.57 0.65 50grad] assert_equals: expected "0.5 - 0.57 0.65 45deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [7 -8 9 400grad] at (0.875) should be [0.5 -0.57 0.65 350grad] assert_equals: expected "0.5 - 0.57 0.65 315deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [7 -8 9 400grad] at (1) should be [0.5 -0.57 0.65 400grad] assert_equals: expected "0.5 - 0.57 0.65 360deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [7 -8 9 400grad] at (2) should be [0.5 -0.57 0.65 800grad] assert_equals: expected "0.5 - 0.57 0.65 720deg " but got "10deg "
PASS CSS Transitions: property <rotate> from [none] to [none] at (-1) should be [none]
PASS CSS Transitions: property <rotate> from [none] to [none] at (0) should be [none]
PASS CSS Transitions: property <rotate> from [none] to [none] at (0.125) should be [none]
PASS CSS Transitions: property <rotate> from [none] to [none] at (0.875) should be [none]
PASS CSS Transitions: property <rotate> from [none] to [none] at (1) should be [none]
PASS CSS Transitions: property <rotate> from [none] to [none] at (2) should be [none]
PASS CSS Transitions with transition: all: property <rotate> from [none] to [none] at (-1) should be [none]
PASS CSS Transitions with transition: all: property <rotate> from [none] to [none] at (0) should be [none]
PASS CSS Transitions with transition: all: property <rotate> from [none] to [none] at (0.125) should be [none]
PASS CSS Transitions with transition: all: property <rotate> from [none] to [none] at (0.875) should be [none]
PASS CSS Transitions with transition: all: property <rotate> from [none] to [none] at (1) should be [none]
PASS CSS Transitions with transition: all: property <rotate> from [none] to [none] at (2) should be [none]
FAIL CSS Animations: property <rotate> from [none] to [none] at (-1) should be [none] assert_equals: expected "none " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [none] at (0) should be [none] assert_equals: expected "none " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [none] at (0.125) should be [none] assert_equals: expected "none " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [none] at (0.875) should be [none] assert_equals: expected "none " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [none] at (1) should be [none] assert_equals: expected "none " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [none] at (2) should be [none] assert_equals: expected "none " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [none] at (-1) should be [none] assert_equals: expected "none " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [none] at (0) should be [none] assert_equals: expected "none " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [none] at (0.125) should be [none] assert_equals: expected "none " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [none] at (0.875) should be [none] assert_equals: expected "none " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [none] at (1) should be [none] assert_equals: expected "none " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [none] at (2) should be [none] assert_equals: expected "none " but got "10deg "
FAIL CSS Transitions: property <rotate> from [none] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from [none] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from [none] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from [none] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "30deg "
PASS CSS Transitions: property <rotate> from [none] to [30deg] at (1) should be [30deg]
FAIL CSS Transitions: property <rotate> from [none] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "30deg "
PASS CSS Transitions with transition: all: property <rotate> from [none] to [30deg] at (1) should be [30deg]
FAIL CSS Transitions with transition: all: property <rotate> from [none] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "30deg "
FAIL CSS Animations: property <rotate> from [none] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [30deg] at (1) should be [30deg] assert_equals: expected "30deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [none] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [30deg] at (1) should be [30deg] assert_equals: expected "30deg " but got "10deg "
FAIL Web Animations: property <rotate> from [none] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from neutral to [30deg] at (-1) should be [-10deg] assert_equals: expected "- 10deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from neutral to [30deg] at (0) should be [10deg] assert_equals: expected "10deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from neutral to [30deg] at (0.25) should be [15deg] assert_equals: expected "15deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from neutral to [30deg] at (0.75) should be [25deg] assert_equals: expected "25deg " but got "30deg "
PASS CSS Transitions: property <rotate> from neutral to [30deg] at (1) should be [30deg]
FAIL CSS Transitions: property <rotate> from neutral to [30deg] at (2) should be [50deg] assert_equals: expected "50deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from neutral to [30deg] at (-1) should be [-10deg] assert_equals: expected "- 10deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from neutral to [30deg] at (0) should be [10deg] assert_equals: expected "10deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from neutral to [30deg] at (0.25) should be [15deg] assert_equals: expected "15deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from neutral to [30deg] at (0.75) should be [25deg] assert_equals: expected "25deg " but got "30deg "
PASS CSS Transitions with transition: all: property <rotate> from neutral to [30deg] at (1) should be [30deg]
FAIL CSS Transitions with transition: all: property <rotate> from neutral to [30deg] at (2) should be [50deg] assert_equals: expected "50deg " but got "30deg "
FAIL CSS Animations: property <rotate> from neutral to [30deg] at (-1) should be [-10deg] assert_equals: expected "- 10deg " but got "10deg "
PASS CSS Animations: property <rotate> from neutral to [30deg] at (0) should be [10deg]
FAIL CSS Animations: property <rotate> from neutral to [30deg] at (0.25) should be [15deg] assert_equals: expected "15deg " but got "10deg "
FAIL CSS Animations: property <rotate> from neutral to [30deg] at (0.75) should be [25deg] assert_equals: expected "25deg " but got "10deg "
FAIL CSS Animations: property <rotate> from neutral to [30deg] at (1) should be [30deg] assert_equals: expected "30deg " but got "10deg "
FAIL CSS Animations: property <rotate> from neutral to [30deg] at (2) should be [50deg] assert_equals: expected "50deg " but got "10deg "
FAIL Web Animations: property <rotate> from neutral to [30deg] at (-1) should be [-10deg] assert_equals: expected "- 10deg " but got "10deg "
PASS Web Animations: property <rotate> from neutral to [30deg] at (0) should be [10deg]
FAIL Web Animations: property <rotate> from neutral to [30deg] at (0.25) should be [15deg] assert_equals: expected "15deg " but got "10deg "
FAIL Web Animations: property <rotate> from neutral to [30deg] at (0.75) should be [25deg] assert_equals: expected "25deg " but got "10deg "
FAIL Web Animations: property <rotate> from neutral to [30deg] at (1) should be [30deg] assert_equals: expected "30deg " but got "10deg "
FAIL Web Animations: property <rotate> from neutral to [30deg] at (2) should be [50deg] assert_equals: expected "50deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [inherit] to [270deg] at (-1) should be [-90deg] assert_equals: expected "- 90deg " but got "270deg "
FAIL CSS Transitions: property <rotate> from [inherit] to [270deg] at (0) should be [90deg] assert_equals: expected "90deg " but got "270deg "
FAIL CSS Transitions: property <rotate> from [inherit] to [270deg] at (0.25) should be [135deg] assert_equals: expected "135deg " but got "270deg "
FAIL CSS Transitions: property <rotate> from [inherit] to [270deg] at (0.75) should be [225deg] assert_equals: expected "225deg " but got "270deg "
PASS CSS Transitions: property <rotate> from [inherit] to [270deg] at (1) should be [270deg]
FAIL CSS Transitions: property <rotate> from [inherit] to [270deg] at (2) should be [450deg] assert_equals: expected "450deg " but got "270deg "
FAIL CSS Transitions with transition: all: property <rotate> from [inherit] to [270deg] at (-1) should be [-90deg] assert_equals: expected "- 90deg " but got "270deg "
FAIL CSS Transitions with transition: all: property <rotate> from [inherit] to [270deg] at (0) should be [90deg] assert_equals: expected "90deg " but got "270deg "
FAIL CSS Transitions with transition: all: property <rotate> from [inherit] to [270deg] at (0.25) should be [135deg] assert_equals: expected "135deg " but got "270deg "
FAIL CSS Transitions with transition: all: property <rotate> from [inherit] to [270deg] at (0.75) should be [225deg] assert_equals: expected "225deg " but got "270deg "
PASS CSS Transitions with transition: all: property <rotate> from [inherit] to [270deg] at (1) should be [270deg]
FAIL CSS Transitions with transition: all: property <rotate> from [inherit] to [270deg] at (2) should be [450deg] assert_equals: expected "450deg " but got "270deg "
FAIL CSS Animations: property <rotate> from [inherit] to [270deg] at (-1) should be [-90deg] assert_equals: expected "- 90deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [inherit] to [270deg] at (0) should be [90deg] assert_equals: expected "90deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [inherit] to [270deg] at (0.25) should be [135deg] assert_equals: expected "135deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [inherit] to [270deg] at (0.75) should be [225deg] assert_equals: expected "225deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [inherit] to [270deg] at (1) should be [270deg] assert_equals: expected "270deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [inherit] to [270deg] at (2) should be [450deg] assert_equals: expected "450deg " but got "10deg "
FAIL Web Animations: property <rotate> from [inherit] to [270deg] at (-1) should be [-90deg] assert_equals: expected "- 90deg " but got "10deg "
FAIL Web Animations: property <rotate> from [inherit] to [270deg] at (0) should be [90deg] assert_equals: expected "90deg " but got "10deg "
FAIL Web Animations: property <rotate> from [inherit] to [270deg] at (0.25) should be [135deg] assert_equals: expected "135deg " but got "10deg "
FAIL Web Animations: property <rotate> from [inherit] to [270deg] at (0.75) should be [225deg] assert_equals: expected "225deg " but got "10deg "
FAIL Web Animations: property <rotate> from [inherit] to [270deg] at (1) should be [270deg] assert_equals: expected "270deg " but got "10deg "
FAIL Web Animations: property <rotate> from [inherit] to [270deg] at (2) should be [450deg] assert_equals: expected "450deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [unset] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from [unset] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from [unset] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "30deg "
FAIL CSS Transitions: property <rotate> from [unset] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "30deg "
PASS CSS Transitions: property <rotate> from [unset] to [30deg] at (1) should be [30deg]
FAIL CSS Transitions: property <rotate> from [unset] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [unset] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [unset] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [unset] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "30deg "
FAIL CSS Transitions with transition: all: property <rotate> from [unset] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "30deg "
PASS CSS Transitions with transition: all: property <rotate> from [unset] to [30deg] at (1) should be [30deg]
FAIL CSS Transitions with transition: all: property <rotate> from [unset] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "30deg "
FAIL CSS Animations: property <rotate> from [unset] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [unset] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [unset] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [unset] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [unset] to [30deg] at (1) should be [30deg] assert_equals: expected "30deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [unset] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "10deg "
FAIL Web Animations: property <rotate> from [unset] to [30deg] at (-1) should be [-30deg] assert_equals: expected "- 30deg " but got "10deg "
FAIL Web Animations: property <rotate> from [unset] to [30deg] at (0) should be [0deg] assert_equals: expected "0deg " but got "10deg "
FAIL Web Animations: property <rotate> from [unset] to [30deg] at (0.25) should be [7.5deg] assert_equals: expected "7.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [unset] to [30deg] at (0.75) should be [22.5deg] assert_equals: expected "22.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [unset] to [30deg] at (1) should be [30deg] assert_equals: expected "30deg " but got "10deg "
FAIL Web Animations: property <rotate> from [unset] to [30deg] at (2) should be [60deg] assert_equals: expected "60deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [100deg] to [-100deg] at (-1) should be [300deg] assert_equals: expected "300deg " but got "- 100deg "
FAIL CSS Transitions: property <rotate> from [100deg] to [-100deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "- 100deg "
FAIL CSS Transitions: property <rotate> from [100deg] to [-100deg] at (0.25) should be [50deg] assert_equals: expected "50deg " but got "- 100deg "
FAIL CSS Transitions: property <rotate> from [100deg] to [-100deg] at (0.75) should be [-50deg] assert_equals: expected "- 50deg " but got "- 100deg "
PASS CSS Transitions: property <rotate> from [100deg] to [-100deg] at (1) should be [-100deg]
FAIL CSS Transitions: property <rotate> from [100deg] to [-100deg] at (2) should be [-300deg] assert_equals: expected "- 300deg " but got "- 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [-100deg] at (-1) should be [300deg] assert_equals: expected "300deg " but got "- 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [-100deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "- 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [-100deg] at (0.25) should be [50deg] assert_equals: expected "50deg " but got "- 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [-100deg] at (0.75) should be [-50deg] assert_equals: expected "- 50deg " but got "- 100deg "
PASS CSS Transitions with transition: all: property <rotate> from [100deg] to [-100deg] at (1) should be [-100deg]
FAIL CSS Transitions with transition: all: property <rotate> from [100deg] to [-100deg] at (2) should be [-300deg] assert_equals: expected "- 300deg " but got "- 100deg "
FAIL CSS Animations: property <rotate> from [100deg] to [-100deg] at (-1) should be [300deg] assert_equals: expected "300deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [-100deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [-100deg] at (0.25) should be [50deg] assert_equals: expected "50deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [-100deg] at (0.75) should be [-50deg] assert_equals: expected "- 50deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [-100deg] at (1) should be [-100deg] assert_equals: expected "- 100deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [100deg] to [-100deg] at (2) should be [-300deg] assert_equals: expected "- 300deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [-100deg] at (-1) should be [300deg] assert_equals: expected "300deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [-100deg] at (0) should be [100deg] assert_equals: expected "100deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [-100deg] at (0.25) should be [50deg] assert_equals: expected "50deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [-100deg] at (0.75) should be [-50deg] assert_equals: expected "- 50deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [-100deg] at (1) should be [-100deg] assert_equals: expected "- 100deg " but got "10deg "
FAIL Web Animations: property <rotate> from [100deg] to [-100deg] at (2) should be [-300deg] assert_equals: expected "- 300deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (-1) should be [0 1 0 300deg] assert_equals: expected "0 1 0 300deg " but got "0 1 0 - 100deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "0 1 0 - 100deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.25) should be [0 1 0 50deg] assert_equals: expected "0 1 0 50deg " but got "0 1 0 - 100deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.75) should be [0 1 0 -50deg] assert_equals: expected "0 1 0 - 50deg " but got "0 1 0 - 100deg "
PASS CSS Transitions: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (1) should be [0 1 0 -100deg]
FAIL CSS Transitions: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (2) should be [0 1 0 -300deg] assert_equals: expected "0 1 0 - 300deg " but got "0 1 0 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (-1) should be [0 1 0 300deg] assert_equals: expected "0 1 0 300deg " but got "0 1 0 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "0 1 0 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.25) should be [0 1 0 50deg] assert_equals: expected "0 1 0 50deg " but got "0 1 0 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.75) should be [0 1 0 -50deg] assert_equals: expected "0 1 0 - 50deg " but got "0 1 0 - 100deg "
PASS CSS Transitions with transition: all: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (1) should be [0 1 0 -100deg]
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (2) should be [0 1 0 -300deg] assert_equals: expected "0 1 0 - 300deg " but got "0 1 0 - 100deg "
FAIL CSS Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (-1) should be [0 1 0 300deg] assert_equals: expected "0 1 0 300deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.25) should be [0 1 0 50deg] assert_equals: expected "0 1 0 50deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.75) should be [0 1 0 -50deg] assert_equals: expected "0 1 0 - 50deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (1) should be [0 1 0 -100deg] assert_equals: expected "0 1 0 - 100deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (2) should be [0 1 0 -300deg] assert_equals: expected "0 1 0 - 300deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (-1) should be [0 1 0 300deg] assert_equals: expected "0 1 0 300deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.25) should be [0 1 0 50deg] assert_equals: expected "0 1 0 50deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (0.75) should be [0 1 0 -50deg] assert_equals: expected "0 1 0 - 50deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (1) should be [0 1 0 -100deg] assert_equals: expected "0 1 0 - 100deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 100deg] to [0 1 0 -100deg] at (2) should be [0 1 0 -300deg] assert_equals: expected "0 1 0 - 300deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (-1) should be [0.22 -0.55 0.8 300deg] assert_equals: expected "0.22 - 0.55 0.8 300deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0) should be [0.22 -0.55 0.8 100deg] assert_equals: expected "0.22 - 0.55 0.8 100deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.25) should be [0.22 -0.55 0.8 50deg] assert_equals: expected "0.22 - 0.55 0.8 50deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.75) should be [0.22 -0.55 0.8 -50deg] assert_equals: expected "0.22 - 0.55 0.8 - 50deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (1) should be [0.22 -0.55 0.8 -100deg] assert_equals: expected "0.22 - 0.55 0.8 - 100deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (2) should be [0.22 -0.55 0.8 -300deg] assert_equals: expected "0.22 - 0.55 0.8 - 300deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (-1) should be [0.22 -0.55 0.8 300deg] assert_equals: expected "0.22 - 0.55 0.8 300deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0) should be [0.22 -0.55 0.8 100deg] assert_equals: expected "0.22 - 0.55 0.8 100deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.25) should be [0.22 -0.55 0.8 50deg] assert_equals: expected "0.22 - 0.55 0.8 50deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.75) should be [0.22 -0.55 0.8 -50deg] assert_equals: expected "0.22 - 0.55 0.8 - 50deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (1) should be [0.22 -0.55 0.8 -100deg] assert_equals: expected "0.22 - 0.55 0.8 - 100deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (2) should be [0.22 -0.55 0.8 -300deg] assert_equals: expected "0.22 - 0.55 0.8 - 300deg " but got "1 - 2.5 3.64 - 100deg "
FAIL CSS Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (-1) should be [0.22 -0.55 0.8 300deg] assert_equals: expected "0.22 - 0.55 0.8 300deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0) should be [0.22 -0.55 0.8 100deg] assert_equals: expected "0.22 - 0.55 0.8 100deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.25) should be [0.22 -0.55 0.8 50deg] assert_equals: expected "0.22 - 0.55 0.8 50deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.75) should be [0.22 -0.55 0.8 -50deg] assert_equals: expected "0.22 - 0.55 0.8 - 50deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (1) should be [0.22 -0.55 0.8 -100deg] assert_equals: expected "0.22 - 0.55 0.8 - 100deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (2) should be [0.22 -0.55 0.8 -300deg] assert_equals: expected "0.22 - 0.55 0.8 - 300deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (-1) should be [0.22 -0.55 0.8 300deg] assert_equals: expected "0.22 - 0.55 0.8 300deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0) should be [0.22 -0.55 0.8 100deg] assert_equals: expected "0.22 - 0.55 0.8 100deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.25) should be [0.22 -0.55 0.8 50deg] assert_equals: expected "0.22 - 0.55 0.8 50deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (0.75) should be [0.22 -0.55 0.8 -50deg] assert_equals: expected "0.22 - 0.55 0.8 - 50deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (1) should be [0.22 -0.55 0.8 -100deg] assert_equals: expected "0.22 - 0.55 0.8 - 100deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 -2.5 3.64 100deg] to [1 -2.5 3.64 -100deg] at (2) should be [0.22 -0.55 0.8 -300deg] assert_equals: expected "0.22 - 0.55 0.8 - 300deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (-1) should be [0 1 0 -10deg] assert_equals: expected "0 1 0 - 10deg " but got "0 1 0 10deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0) should be [0 1 0 0deg] assert_equals: expected "0 1 0 0deg " but got "0 1 0 10deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.25) should be [0 1 0 2.5deg] assert_equals: expected "0 1 0 2.5deg " but got "0 1 0 10deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.75) should be [0 1 0 7.5deg] assert_equals: expected "0 1 0 7.5deg " but got "0 1 0 10deg "
PASS CSS Transitions: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (1) should be [0 1 0 10deg]
FAIL CSS Transitions: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (2) should be [0 1 0 20deg] assert_equals: expected "0 1 0 20deg " but got "0 1 0 10deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (-1) should be [0 1 0 -10deg] assert_equals: expected "0 1 0 - 10deg " but got "0 1 0 10deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0) should be [0 1 0 0deg] assert_equals: expected "0 1 0 0deg " but got "0 1 0 10deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.25) should be [0 1 0 2.5deg] assert_equals: expected "0 1 0 2.5deg " but got "0 1 0 10deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.75) should be [0 1 0 7.5deg] assert_equals: expected "0 1 0 7.5deg " but got "0 1 0 10deg "
PASS CSS Transitions with transition: all: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (1) should be [0 1 0 10deg]
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (2) should be [0 1 0 20deg] assert_equals: expected "0 1 0 20deg " but got "0 1 0 10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (-1) should be [0 1 0 -10deg] assert_equals: expected "0 1 0 - 10deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0) should be [0 1 0 0deg] assert_equals: expected "0 1 0 0deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.25) should be [0 1 0 2.5deg] assert_equals: expected "0 1 0 2.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.75) should be [0 1 0 7.5deg] assert_equals: expected "0 1 0 7.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (1) should be [0 1 0 10deg] assert_equals: expected "0 1 0 10deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (2) should be [0 1 0 20deg] assert_equals: expected "0 1 0 20deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (-1) should be [0 1 0 -10deg] assert_equals: expected "0 1 0 - 10deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0) should be [0 1 0 0deg] assert_equals: expected "0 1 0 0deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.25) should be [0 1 0 2.5deg] assert_equals: expected "0 1 0 2.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (0.75) should be [0 1 0 7.5deg] assert_equals: expected "0 1 0 7.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (1) should be [0 1 0 10deg] assert_equals: expected "0 1 0 10deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 0deg] to [0 1 0 10deg] at (2) should be [0 1 0 20deg] assert_equals: expected "0 1 0 20deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (-1) should be [0.67 -0.06 -0.74 124.97deg] assert_equals: expected "0.67 - 0.06 - 0.74 124.97deg " but got "0 1 1 135deg "
FAIL CSS Transitions: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0) should be [0.71 0.71 0 90deg] assert_equals: expected "0.71 0.71 0 90deg " but got "0 1 1 135deg "
FAIL CSS Transitions: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.25) should be [0.54 0.8 0.26 94.83deg] assert_equals: expected "0.54 0.8 0.26 94.83deg " but got "0 1 1 135deg "
FAIL CSS Transitions: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.75) should be [0.17 0.78 0.61 118.68deg] assert_equals: expected "0.17 0.78 0.61 118.68deg " but got "0 1 1 135deg "
FAIL CSS Transitions: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (1) should be [0 0.71 0.71 135deg] assert_equals: expected "0 0.71 0.71 135deg " but got "0 1 1 135deg "
FAIL CSS Transitions: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (2) should be [-0.52 0.29 0.81 208.96deg] assert_equals: expected "- 0.52 0.29 0.81 208.96deg " but got "0 1 1 135deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (-1) should be [0.67 -0.06 -0.74 124.97deg] assert_equals: expected "0.67 - 0.06 - 0.74 124.97deg " but got "0 1 1 135deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0) should be [0.71 0.71 0 90deg] assert_equals: expected "0.71 0.71 0 90deg " but got "0 1 1 135deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.25) should be [0.54 0.8 0.26 94.83deg] assert_equals: expected "0.54 0.8 0.26 94.83deg " but got "0 1 1 135deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.75) should be [0.17 0.78 0.61 118.68deg] assert_equals: expected "0.17 0.78 0.61 118.68deg " but got "0 1 1 135deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (1) should be [0 0.71 0.71 135deg] assert_equals: expected "0 0.71 0.71 135deg " but got "0 1 1 135deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (2) should be [-0.52 0.29 0.81 208.96deg] assert_equals: expected "- 0.52 0.29 0.81 208.96deg " but got "0 1 1 135deg "
FAIL CSS Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (-1) should be [0.67 -0.06 -0.74 124.97deg] assert_equals: expected "0.67 - 0.06 - 0.74 124.97deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0) should be [0.71 0.71 0 90deg] assert_equals: expected "0.71 0.71 0 90deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.25) should be [0.54 0.8 0.26 94.83deg] assert_equals: expected "0.54 0.8 0.26 94.83deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.75) should be [0.17 0.78 0.61 118.68deg] assert_equals: expected "0.17 0.78 0.61 118.68deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (1) should be [0 0.71 0.71 135deg] assert_equals: expected "0 0.71 0.71 135deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (2) should be [-0.52 0.29 0.81 208.96deg] assert_equals: expected "- 0.52 0.29 0.81 208.96deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (-1) should be [0.67 -0.06 -0.74 124.97deg] assert_equals: expected "0.67 - 0.06 - 0.74 124.97deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0) should be [0.71 0.71 0 90deg] assert_equals: expected "0.71 0.71 0 90deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.25) should be [0.54 0.8 0.26 94.83deg] assert_equals: expected "0.54 0.8 0.26 94.83deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (0.75) should be [0.17 0.78 0.61 118.68deg] assert_equals: expected "0.17 0.78 0.61 118.68deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (1) should be [0 0.71 0.71 135deg] assert_equals: expected "0 0.71 0.71 135deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 1 0 90deg] to [0 1 1 135deg] at (2) should be [-0.52 0.29 0.81 208.96deg] assert_equals: expected "- 0.52 0.29 0.81 208.96deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (-1) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "1 0 0 450deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "1 0 0 450deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.25) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "1 0 0 450deg "
FAIL CSS Transitions: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.75) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "1 0 0 450deg "
PASS CSS Transitions: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (1) should be [1 0 0 450deg]
FAIL CSS Transitions: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (2) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "1 0 0 450deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (-1) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "1 0 0 450deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "1 0 0 450deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.25) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "1 0 0 450deg "
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.75) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "1 0 0 450deg "
PASS CSS Transitions with transition: all: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (1) should be [1 0 0 450deg]
FAIL CSS Transitions with transition: all: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (2) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "1 0 0 450deg "
FAIL CSS Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (-1) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.25) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.75) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (1) should be [1 0 0 450deg] assert_equals: expected "1 0 0 450deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (2) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (-1) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.25) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (0.75) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (1) should be [1 0 0 450deg] assert_equals: expected "1 0 0 450deg " but got "10deg "
FAIL Web Animations: property <rotate> from [0 1 0 0deg] to [1 0 0 450deg] at (2) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "10deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (-1) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "0 1 0 0deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0) should be [1 0 0 450deg] assert_equals: expected "1 0 0 450deg " but got "0 1 0 0deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.25) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "0 1 0 0deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.75) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "0 1 0 0deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (1) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "0 1 0 0deg "
FAIL CSS Transitions: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (2) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "0 1 0 0deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (-1) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "0 1 0 0deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0) should be [1 0 0 450deg] assert_equals: expected "1 0 0 450deg " but got "0 1 0 0deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.25) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "0 1 0 0deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.75) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "0 1 0 0deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (1) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "0 1 0 0deg "
FAIL CSS Transitions with transition: all: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (2) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "0 1 0 0deg "
FAIL CSS Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (-1) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0) should be [1 0 0 450deg] assert_equals: expected "1 0 0 450deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.25) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.75) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (1) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "10deg "
FAIL CSS Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (2) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (-1) should be [1 0 0 900deg] assert_equals: expected "1 0 0 900deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0) should be [1 0 0 450deg] assert_equals: expected "1 0 0 450deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.25) should be [1 0 0 337.5deg] assert_equals: expected "1 0 0 337.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (0.75) should be [1 0 0 112.5deg] assert_equals: expected "1 0 0 112.5deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (1) should be [1 0 0 0deg] assert_equals: expected "1 0 0 0deg " but got "10deg "
FAIL Web Animations: property <rotate> from [1 0 0 450deg] to [0 1 0 0deg] at (2) should be [1 0 0 -450deg] assert_equals: expected "1 0 0 - 450deg " but got "10deg "
